requirements.txt
python-dotenv~=1.0.1
SQLAlchemy~=2.0.30
redis~=5.0.5
aiogram~=3.7.0
jdatetime~=5.0.0
aiofiles~=23.2.1
aiosqlite~=0.20.0
aiohttp-socks~=0.8.4
APScheduler~=3.10.4